requirements.txt
pandas==2.3.2
scikit-learn==1.7.2
ipykernel==6.30.1
ipython==9.5.0
python-telegram-bot==20.*